+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `Response::blob`
 - Added `panic-hook` feature, enabled by default, to conditionally include `console_error_panic_hook`
 - [BREAKING] Hid markdown funnctionality behind optional `markdown` feature
+- Implemented `AsAtValue` for `Option<T>`
 
 ## v0.8.0
 - [BREAKING] Rename `linear_gradient!` to `linearGradient!` for consistency with the other svg macros (same with `radial_gradient!` and `mesh_gradient!`) (#377).